test: Created Performance Test
establishes the baseline for the detect_prey by running it 100,000 times. Looks like the issue is in `distance` of the 18 seconds in `detect_prey` 14 of those seconds are in `distance`. So the goal is to reduce the calls to distance, or speed up `distance`.

Baseline: (18.261 seconds)
   100000    3.770    0.000   18.261    0.000 agents.py:105(detect_prey)
  9900000   10.998    0.000   14.013    0.000 agents.py:16(distance)

Note: THIS TEST IS NOT A GOOD TEST AND IS JUST HERE TO ESTABLISH A BASELINE. YOU CAN DELETE THIS LATER IF IT STINKS