performance: Reduced calls to distance
`max` is the next slowest thing in `detect_prey`. Realized that the max is not needed here but in the one other use of the function. so moved it there.
Baseline: (11.668 seconds)
   100000    4.014    0.000   11.668    0.000 agents.py:105(detect_prey)
  2500681    0.420    0.000    0.420    0.000 {built-in method builtins.max}
After: (9.915 seconds)
   100000    4.329    0.000    9.915    0.000 agents.py:105(detect_prey)

Note: NOT sure what happened here but the time DROPPED A LOT. might be my system.