performance: replaced deepcopy with new object creation
deepcopy is convent, but slow and often hides the details of what is going on.

Baseline: (14.016 seconds)
   100000    0.256    0.000   14.016    0.000 agents.py:168(find_partner)
   100000    3.808    0.000    7.771    0.000 agents.py:105(detect_prey)
2850000/150000    2.728    0.000    5.328    0.000 copy.py:132(deepcopy)
After: (9.232 seconds)
   100000    0.190    0.000    9.232    0.000 agents.py:168(find_partner)
   100000    3.868    0.000    8.434    0.000 agents.py:105(detect_prey)
   deepcopy NOT CALLED

Note: Based on the reproduction work here there is no mutation happening here, just selection of the fastest one. maybe vary the speed by a random value and/or select a random parent speed plus a rand value? Just an idea.

 NOTE: THIS IS JUST BREAKING OUT THE EXISTING CODE AND NOT RESTRUCTURING TO BE MORE READABLE. WHATEVER READABILITY GAINED IS ACCIDENTAL.