performance: `random_movement` not slowest, but noticed the recursion as not optimal.
By popping the move off the list and passing the list in the recursion each recursive step has less moves to choice from and thus less recursion, the current implementation that you have allows for an infinite recursion to happen (unlikely but possible).

Notice the reduced number of calls from 10289 to 3701.

Baseline: (0.040 seconds)
10289/10112    0.009    0.000    0.040    0.000 agents.py:90(random_movement)

After: (0.017 seconds)
3701/3640    0.005    0.000    0.017    0.000 agents.py:90(random_movement)